Delete validations

+++ src/main/java/com/jsfmrydvd/home/TodoDelete.java
@@ -16,7 +16,15 @@ public class TodoDelete extends HttpServlet {
 
     protected void doGet(HttpServletRequest request, HttpServletResponse response)
             throws IOException, ServletException {
-        homeService.delete(new Todo(request.getParameter("todo")));
-        response.sendRedirect("/home");
+        String names = (String) request.getSession().getAttribute("name");
+        if(request.getParameter("todo").contains(names)) {
+            homeService.delete(new Todo(request.getParameter("todo")));
+            response.sendRedirect("/home");
+        } else {
+            request.setAttribute("errorMessage", "You can only delete your own message!");
+            request.setAttribute("todos", homeService.retrieveTodos());
+            request.getRequestDispatcher("/WEB-INF/views/home.jsp").forward(request, response);
+        }
+
     }
 }
